requirements.txt
transformers
torch
beautifulsoup4
python-dotenv
imapclient
mail-parser
dateparser
sqlalchemy
requests
notion-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
schedule


